Animation Testing
This push includes the AnimationTesting scene, which has a very rudimentary NPC face. Running the scene will show off the character blinking, and pressing the Up arrow will toggle an angry expression.

I don't fully understand Animation transitions in Unity, and there are several obvious errors in the animation, but it's a start.